% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_3}
\alias{test_3}
\title{Descriptive data for BRMS debugging}
\format{
An object of class \code{list} of length 3.
}
\usage{
test_3
}
\description{
General model formula:

```
f0 <- brms::brmsformula(
  respons ~ pnt_prd + tm_snc + SGU + z_bins_fac2 + fSCA + prcp + tmin + tmax +
  sar(M = w0, type = 'error') +
  ar(time = year, gr = pixel, p = 1),
  family = gaussian()
)
```

Model run:

```
mod0 <- brms::brm(
  formula = f0, data = cb_df, data2 = list(w0 = w0),
  prior = NULL, iter = 2000, cores = 4, chains = 4,
  silent = 0
)
```

The first element of the list is the input data.frame, for `test_3` this is a `data.frame` with 1480 rows and 33 columns.

This model ran successfully.
}
\keyword{datasets}
